requirements.txt
scipy==1.5.4
dataframe_image==0.1.1
networkx==2.5.1
goatools==1.1.5
node2vec==0.4.3
matplotlib==3.4.1
shap==0.39.0
statsmodels==0.12.2
tqdm==4.30.0
pandas==1.0.5
numpy==1.19.1
imbalanced_learn==0.8.0
xgboost==1.1.1
imblearn==0.0
python_igraph==0.9.6
scikit_learn==1.0
